LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/animation-events-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: ViewTimeline

Harness Error (FAIL), message = ReferenceError: Can't find variable: ViewTimeline
